backend/requirements.txt
flask
flask-cors
youtube-transcript-api
google-generativeai
python-dotenv
yt-dlp
openai-whisper
gunicorn
